feat: extract list to new ReleasedVersionsList component
