+++ frontend/src/components/ReleasedVersionsList/ReleasedVersionsList.module.css
@@ -1,0 +1,22 @@
+.version-list-header {
+  display: grid;
+  grid-template-columns: 150px 150px 1fr 1fr;
+  padding: 10px 0;
+  font-weight: bold;
+  border-bottom: 2px solid #eaeaea;
+  margin-bottom: 10px;
+}
+
+.version-list {
+  list-style-type: none;
+  padding: 0;
+  margin: 0;
+}
+
+.version-item {
+  display: grid;
+  grid-template-columns: 150px 150px 1fr 1fr;
+  padding: 12px 0;
+  border-bottom: 1px solid #f0f0f0;
+  align-items: center;
+}
